requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
djangorestframework==3.12.1
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
joblib==0.17.0
numpy==1.19.3
psycopg2==2.8.6
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.3
sqlparse==0.4.1
static3==0.7.0
threadpoolctl==2.1.0
whitenoise==5.2.0
